Refactored: Embeddings and Parsings
