+++ src/models/model.config.ts
@@ -1,6 +1,7 @@
 export interface LocalConfig {
   projectName: string;
   embedding: boolean;
+  depgraph: boolean;
   ignore: string[];
 }
